# openapi/customer-api.yaml
openapi: 3.0.3
info:
  title: Customer Information API (WireMock proxy to Excel/CSV backend)
  version: 1.0.0
servers:
  - url: http://localhost:8080
paths:
  /customers/{customerId}:
    get:
      summary: Get customer by ID
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Customer not found in database
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customers:
    post:
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Created
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCustomerResponse"
components:
  schemas:
    Customer:
      type: object
      properties:
        customerId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        status:
          type: string
        registeredAt:
          type: string
          format: date-time
    CreateCustomerRequest:
      type: object
      required: [firstName, lastName, email]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
    CreateCustomerResponse:
      type: object
      properties:
        customerId:
          type: string
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string